añadida a sandbox señal global para timeout

+++ rendu4/sandbox/sandbox.c
@@ -39,7 +39,7 @@ int sandbox(void (*f)(void), unsigned int timeout, bool verbose)
     int ret = waitpid(pid, &status, 0);
     alarm(0); //cancela si termina antes
 
-    if(ret == -1 && errno == EINTR) //timeout
+    if(ret == -1 && errno == EINTR && g_timeout) //timeout
     {
         kill(pid, SIGKILL); //matar hijo
         waitpid(pid, NULL, 0); //recojo zombie
